requirements_base.txt
dj-database-url==0.2.2
gunicorn==18.0
static==1.0.2
otree-core==0.3.14
boto==2.38.0
django-s3direct==0.3.11
psycopg2==2.5.1
django-toolbelt==0.0.1

